# docs/swagger.yaml
definitions:
  user.User:
    properties:
      address:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phoneNumber:
        type: string
    type: object
host: localhost:12000
info:
  contact:
    email: ahnus123@gmail.com
    name: ahnus123
    url: https://github.com/ahnus123/go-study/user
  description: This is api document for user server.
  license:
    name: gorilla/mux
    url: https://github.com/gorilla/mux/blob/master/LICENSE
  title: Swagger - User API
  version: "1.0"
paths:
  /user:
    get:
      consumes:
      - application/json
      description: user 목록 조회
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: GetUserList
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: user 수정
      parameters:
      - description: ex) 7f11671a-eeb5-4d0e-bcb0-f83260a6d375
        in: path
        name: id
        required: true
        type: string
      - description: body params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: string
            - properties:
                id:
                  type: string
              type: object
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: UpdateUser
      tags:
      - user
    post:
      consumes:
      - application/json
      description: user 생성
      parameters:
      - description: ex) 7f11671a-eeb5-4d0e-bcb0-f83260a6d375
        in: path
        name: id
        required: true
        type: string
      - description: body params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: string
            - properties:
                id:
                  type: string
              type: object
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: CreateUser
      tags:
      - user
  /user/{userId}:
    delete:
      consumes:
      - application/json
      description: user 삭제
      parameters:
      - description: ex) 7f11671a-eeb5-4d0e-bcb0-f83260a6d375
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: string
            - properties:
                id:
                  type: string
              type: object
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: DeleteUser
      tags:
      - user
    get:
      consumes:
      - application/json
      description: user 조회 by id
      parameters:
      - description: ex) 7f11671a-eeb5-4d0e-bcb0-f83260a6d375
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: GetUser
      tags:
      - user
swagger: "2.0"